+++ wfe-service/src/main/java/ru/runa/wfe/service/ExecutionService.java
@@ -46,19 +46,29 @@ import ru.runa.wfe.var.file.FileVariableImpl;
  * @since 4.0
  */
 public interface ExecutionService {
-    
-    //код Румянцева чат
+
+    // код Румянцева чат
     public List<ChatMessage> getChatMessages(int chatId);
-    public ChatMessage getChatMessage(int chatId,long messageId);
-    public List<ChatMessage> getChatMessages(int chatId,int firstId, int count);
+
+    public ChatMessage getChatMessage(int chatId, long messageId);
+
+    public List<ChatMessage> getChatMessages(int chatId, int firstId, int count);
+
     public List<ChatMessage> getChatFirstMessages(int chatId, int count);
-    //возвращает id нового сообщения в БД
-    public int setChatMessage(int chatId, ChatMessage message);
+
+    // возвращает id нового сообщения в БД
+    public long setChatMessage(int chatId, ChatMessage message);
+
     long getChatAllMessagesCount(int chatId);
+
     public void deleteChatMessage(long messId);
+
     public List<Integer> getChatAllConnectedChatId(int chatId);
-    public ChatsUserInfo getChatUserInfo(long userId,String userName, int chatId);
+
+    public ChatsUserInfo getChatUserInfo(long userId, String userName, int chatId);
+
     public long getChatNewMessagesCount(long lastMessageId, int chatId);
+
     public void updateChatUserInfo(long userId, String userName, int chatId, long lastMessageId);
     //
 
@@ -75,8 +85,8 @@ public interface ExecutionService {
      * @throws DefinitionDoesNotExistException
      * @throws ValidationException
      */
-    public Long startProcess(User user, String definitionName, Map<String, Object> variables) throws DefinitionDoesNotExistException,
-            ValidationException;
+    public Long startProcess(User user, String definitionName, Map<String, Object> variables)
+            throws DefinitionDoesNotExistException, ValidationException;
 
     /**
      * Starts new process by definition.
@@ -91,8 +101,8 @@ public interface ExecutionService {
      * @throws DefinitionDoesNotExistException
      * @throws ValidationException
      */
-    public Long startProcessById(User user, Long definitionId, Map<String, Object> variables) throws DefinitionDoesNotExistException,
-            ValidationException;
+    public Long startProcessById(User user, Long definitionId, Map<String, Object> variables)
+            throws DefinitionDoesNotExistException, ValidationException;
 
     /**
      * Gets process count for {@link BatchPresentation}.
@@ -186,7 +196,7 @@ public interface ExecutionService {
      * @throws ProcessDoesNotExistException
      */
     public List<WfSwimlane> getActiveProcessesSwimlanes(User user, String namePattern);
-    
+
     /**
      * Reassigns initialized process role.
      *
@@ -198,7 +208,7 @@ public interface ExecutionService {
      * @throws ProcessDoesNotExistException
      */
     public boolean reassignSwimlane(User user, Long id);
-    
+
     /**
      * Assigns role by name to specified executor.
      *
